Move platform activation to shared header
