+++ dev/AppNotifications/AppNotificationManager.cpp
@@ -26,6 +26,7 @@
 #include <Microsoft.RoApi.h>
 #include <ShellLocalization.h>
 #include <filesystem>
+#include <NotificationPlatformActivation.h>
 
 using namespace std::literals;
 
@@ -228,7 +229,7 @@ namespace winrt::Microsoft::Windows::AppNotifications::implementation
 
     void AppNotificationManager::RegisterAppNotificationSinkWithLongRunningPlatform()
     {
-        auto notificationPlatform{ PushNotificationHelpers::GetNotificationPlatform() };
+        auto notificationPlatform{ NotificationPlatform::GetNotificationPlatform() };
         THROW_IF_FAILED(notificationPlatform->AddToastRegistrationMapping(m_processName.c_str(), m_appId.c_str()));
     }
 
@@ -332,7 +333,7 @@ namespace winrt::Microsoft::Windows::AppNotifications::implementation
             // Remove any Registrations from the Long Running Process that are necessary for Cloud toasts
             if (!PushNotificationHelpers::IsPackagedAppScenario() && !WindowsAppRuntime::SelfContained::IsSelfContained())
             {
-                auto notificationPlatform{ PushNotificationHelpers::GetNotificationPlatform() };
+                auto notificationPlatform{ NotificationPlatform::GetNotificationPlatform() };
                 THROW_IF_FAILED(notificationPlatform->RemoveToastRegistrationMapping(m_processName.c_str()));
             }
