+++ dev/PushNotifications/PushNotificationManager.cpp
@@ -23,6 +23,7 @@
 #include <Microsoft.RoApi.h>
 #include "PushNotificationReceivedEventArgs.h"
 #include <security.integritylevel.h>
+#include <NotificationPlatformActivation.h>
 
 using namespace std::literals;
 using namespace Microsoft::Windows::AppNotifications::Helpers;
@@ -273,7 +274,7 @@ namespace winrt::Microsoft::Windows::PushNotifications::implementation
                     }
                     else
                     {
-                        auto notificationPlatform{ PushNotificationHelpers::GetNotificationPlatform() };
+                        auto notificationPlatform{ NotificationPlatform::GetNotificationPlatform() };
 
                         // AppId is generated by PushNotificationLongRunningTask singleton
                         wil::unique_cotaskmem_string unpackagedAppUserModelId;
@@ -526,7 +527,7 @@ namespace winrt::Microsoft::Windows::PushNotifications::implementation
                     }
                 }) };
 
-                auto notificationPlatform{ PushNotificationHelpers::GetNotificationPlatform() };
+                auto notificationPlatform{ NotificationPlatform::GetNotificationPlatform() };
                 wil::unique_cotaskmem_string unpackagedAppUserModelId;
 
                 std::wstring processName;
@@ -542,7 +543,7 @@ namespace winrt::Microsoft::Windows::PushNotifications::implementation
                 }) };
 
                 // Registers a NoticicationListener and foreground delegate with the Long Running Process singleton
-                auto notificationsLongRunningPlatform{ PushNotificationHelpers::GetNotificationPlatform() };
+                auto notificationsLongRunningPlatform{ NotificationPlatform::GetNotificationPlatform() };
 
                 // Register a sink with platform brokered through PushNotificationsLongRunningProcess
                 THROW_IF_FAILED(notificationsLongRunningPlatform->RegisterForegroundActivator(PushNotificationManager::Default().as<IWpnForegroundSink>().get(), processName.c_str()));
@@ -672,7 +673,7 @@ namespace winrt::Microsoft::Windows::PushNotifications::implementation
                     processName = m_processName;
                 }
 
-                auto notificationsLongRunningPlatform{ PushNotificationHelpers::GetNotificationPlatform() };
+                auto notificationsLongRunningPlatform{ NotificationPlatform::GetNotificationPlatform() };
                 // Unregister foreground sink with the Long Running Process Singleton
                 THROW_IF_FAILED(notificationsLongRunningPlatform->UnregisterForegroundActivator(processName.c_str()));
 
@@ -755,7 +756,7 @@ namespace winrt::Microsoft::Windows::PushNotifications::implementation
                     processName = m_processName;
                 }
 
-                auto notificationsLongRunningPlatform{ PushNotificationHelpers::GetNotificationPlatform() };
+                auto notificationsLongRunningPlatform{ NotificationPlatform::GetNotificationPlatform() };
                 // Removes the Long Running Singleton sink registered with platform for both packaged and unpackaged applications 
                 THROW_IF_FAILED(notificationsLongRunningPlatform->UnregisterLongRunningActivator(processName.c_str()));
