+++ dev/PushNotifications/PushNotificationUtility.h
@@ -128,23 +128,6 @@ namespace winrt::Microsoft::Windows::PushNotifications::Helpers
     }
     CATCH_RETURN()
 
-    inline wil::com_ptr<INotificationsLongRunningPlatform> GetNotificationPlatform()
-    {
-        // If supported, allow use of class registrations from less trusted sources.
-        auto supportMarker{ wil::CoCreateInstanceNoThrow<ISupportAllowLowerTrustActivation>(CLSID_ActivationCapabilities) };
-        if (supportMarker)
-        {
-            return wil::CoCreateInstance<INotificationsLongRunningPlatform>(CLSID_NotificationsLongRunningPlatform, CLSCTX_LOCAL_SERVER |
-                CLSCTX_ALLOW_LOWER_TRUST_REGISTRATION);
-        }
-        else
-        {
-            // Even if we can't pass CLSCTX_ALLOW_LOWER_TRUST_REGISTRATION, we still want to attempt the activation.
-            // This might fail to find the class when run elevated on a downlevel SKU, but it will work otherwise.
-            return wil::CoCreateInstance<INotificationsLongRunningPlatform>(CLSID_NotificationsLongRunningPlatform, CLSCTX_LOCAL_SERVER);
-        }
-    }
-
     inline bool IsBackgroundTaskBuilderAvailable()
     {
         static bool hasSetTaskEntrypoint{ winrt::Metadata::ApiInformation::IsMethodPresent(L"Windows.ApplicationModel.Background.BackgroundTaskBuilder", L"SetTaskEntryPointClsid") };
