+++ dev/PushNotifications/PushNotificationsLongRunningTask.StartupTask/PushNotificationsLongRunningPlatform-Startup.h
@@ -6,6 +6,7 @@
 #include <wil/com.h>
 #include <wil/resource.h>
 #include "../../Common/Microsoft.RoApi.h"
+#include "../../PushNotifications/NotificationPlatformActivation.h"
 
 inline bool isRetriableRpcError(HRESULT hr)
 {
@@ -26,20 +27,7 @@ inline HRESULT StartupNotificationsLongRunningPlatform()
         try
         {
             ::Microsoft::RoApi::ClearRoActivateInstanceCache();
-
-            // If supported, allow use of class registrations from less trusted sources.
-            auto supportMarker{ wil::CoCreateInstanceNoThrow<ISupportAllowLowerTrustActivation>(CLSID_ActivationCapabilities) };
-            if (supportMarker)
-            {
-                wil::CoCreateInstance<INotificationsLongRunningPlatform>(CLSID_NotificationsLongRunningPlatform, CLSCTX_LOCAL_SERVER |
-                    CLSCTX_ALLOW_LOWER_TRUST_REGISTRATION);
-            }
-            else
-            {
-                // Even if we can't pass CLSCTX_ALLOW_LOWER_TRUST_REGISTRATION, we still want to attempt the activation.
-                // This might fail to find the class when run elevated on a downlevel SKU, but it will work otherwise.
-                wil::CoCreateInstance<INotificationsLongRunningPlatform>(CLSID_NotificationsLongRunningPlatform, CLSCTX_LOCAL_SERVER);
-            }
+            NotificationPlatform::GetNotificationPlatform();
             break;
         }
         catch (...)
